Moves ListView to its own module
ListView file was getting somewhat large, I wanted to split it up into separate files to make it easier to find the applicable datatype.
