+++ mps_youtube/listview/songtitle.py
@@ -1,0 +1,29 @@
+from .base import ListViewItem
+from .. import util as u
+
+
+class ListSongtitle(ListViewItem):
+    """ Describes a user
+    """
+    # pylint: disable=unused-argument
+    _checked = False
+
+    def artist(self, l=10):
+        """ Get artist """
+        return u.uea_pad(l, self.data[0])
+
+    def title(self, l=10):
+        """ Get title """
+        return u.uea_pad(l, self.data[1])
+
+    def checked(self, l=10):
+        """ String from for checked """
+        return "  X  " if self.checked() else "     "
+
+    def is_checked(self):
+        """ Returns true if checked """
+        return self._checked
+
+    def toggle(self):
+        """ Toggle checked status """
+        self._checked = not self._checked
